refactor(e2e): signUp
